+++ src/main/groovy/nextflow/validation/config/LoggingConfig.groovy
@@ -21,23 +21,39 @@ class LoggingConfig implements ConfigScope {
 
     @ConfigOption
     @Description('Define the logging level of unrecognised parameters. Defaults to `warn`.')
-    final public ValidationLogger logUnrecognisedParams
+    final public ValidationLogger unrecognisedParams
+
+    @ConfigOption
+    @Description('Define the logging level of unrecognised headers that are found in the samplesheets. Defaults to `warn`.')
+    final public ValidationLogger unrecognisedHeaders
 
     final private OPTIONS = ['skip', 'debug', 'info', 'warn', 'error']
 
-    LoggingConfig(Map map){
+    LoggingConfig(Map map, Boolean monochromeLogs = false) {
         def config = map ?: [:]
 
-        // logUnrecognisedParams
-        def String level = config.get("logUnrecognisedParams", "warn")
+        // unrecognisedParams
+        def String level = config.get("unrecognisedParams", "warn")
         if(OPTIONS.contains(level)) {
-            if(config.get("logUnrecognisedParams")) {
-                log.debug("Set `validation.logUnrecognisedParams` to ${level}")
+            if(config.get("unrecognisedParams")) {
+                log.debug("Set `validation.unrecognisedParams` to ${level}")
             }
-            logUnrecognisedParams = new ValidationLogger(level)
+            unrecognisedParams = new ValidationLogger(level, monochromeLogs)
         } else {
-            log.warn("Incorrect value detected for `validation.logUnrecognisedParams`, one of (${OPTIONS.join(', ')}) is expected. Defaulting to `warn`")
-            logUnrecognisedParams = new ValidationLogger("warn")
+            log.warn("Incorrect value detected for `validation.unrecognisedParams`, one of (${OPTIONS.join(', ')}) is expected. Defaulting to `warn`")
+            unrecognisedParams = new ValidationLogger("warn", monochromeLogs)
+        }
+
+        // unrecognisedHeaders
+        level = config.get("unrecognisedHeaders", "warn")
+        if(OPTIONS.contains(level)) {
+            if(config.get("unrecognisedHeaders")) {
+                log.debug("Set `validation.unrecognisedHeaders` to ${level}")
+            }
+            unrecognisedHeaders = new ValidationLogger(level, monochromeLogs)
+        } else {
+            log.warn("Incorrect value detected for `validation.unrecognisedHeaders`, one of (${OPTIONS.join(', ')}) is expected. Defaulting to `warn`")
+            unrecognisedHeaders = new ValidationLogger("warn", monochromeLogs)
         }
     }
 }