+++ src/main/groovy/nextflow/validation/logging/ValidationLogger.groovy
@@ -3,6 +3,8 @@ package nextflow.validation.logging
 import groovy.util.logging.Slf4j
 
 import nextflow.validation.exceptions.SchemaValidationException
+import static nextflow.validation.utils.Colors.getLogColors
+
 
 /**
  * This class is used to define logging logic that can be configured via configuration options.
@@ -14,8 +16,9 @@ import nextflow.validation.exceptions.SchemaValidationException
 @Slf4j
 class ValidationLogger {
     final private String level = 'info'
+    final private Map colors
 
-    ValidationLogger(String level) {
+    ValidationLogger(String level, Boolean monochromeLogs = false) {
         level = level?.toLowerCase()
         if (level in ['skip', 'debug', 'info', 'warn', 'error']) {
             this.level = level
@@ -23,6 +26,7 @@ class ValidationLogger {
             log.error("Invalid log level '${level}' specified, defaulting to 'info'")
             System.exit 1
         }
+        this.colors = getLogColors(monochromeLogs)
     }
 
     public void log(String message) {
@@ -40,7 +44,7 @@ class ValidationLogger {
                 log.warn(message)
                 break
             case 'error':
-                throw new SchemaValidationException(message)
+                throw new SchemaValidationException(colors.red + message + colors.reset)
                 break
             default:
                 log.info(message) // Fallback to info if something goes wrong