+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Changes
 
 1. The plugin now properly validates cloud storage files instead of skipping them. Exotic errors will be added to the error messages instead of failing the validation outright.
+2. Migrated to the new observer class in Nextflow 25.04.0
 
 # Version 2.4.2